+++ Backend/src/routes/controllers/getAllEventsController.js
@@ -1,0 +1,8 @@
+const { Events } = require("../../db")
+
+const getAllEvents = async()=>{
+    const allEvents = await Events.findAll();
+    return allEvents;
+}
+
+module.exports = {getAllEvents}
